FBP-61. Separate package for event-alike elements and added class stubs
