--- xml-parser-api/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/bpmn/api/bpmn/elements/BpmnEvent.kt
@@ -1,7 +1,0 @@
-package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements
-
-import com.github.pozo.KotlinBuilder
-import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.BpmnElementId
-
-@KotlinBuilder
-data class BpmnEvent(val id: BpmnElementId) {}
