+++ xml-parser-api/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/bpmn/api/bpmn/elements/events/end/BpmnEndErrorEvent.kt
@@ -1,0 +1,2 @@
+package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.events.end
+
