+++ xml-parser-api/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/bpmn/api/bpmn/elements/events/end/BpmnEndEvent.kt
@@ -1,7 +1,8 @@
-package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements
+package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.events.end
 
 import com.github.pozo.KotlinBuilder
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.BpmnElementId
+import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.WithBpmnId
 
 @KotlinBuilder
 data class BpmnEndEvent(
