+++ xml-parser-api/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/bpmn/api/bpmn/elements/gateways/BpmnInclusiveGateway.kt
@@ -1,0 +1,4 @@
+package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.gateways
+
+class BpmnInclusiveGateway {
+}
