+++ xml-parser-api/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/bpmn/api/bpmn/elements/subprocess/BpmnTransactionalSubprocess.kt
@@ -1,0 +1,3 @@
+package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.subprocess
+
+class BpmnTransactionalSubprocess
